Fix get dependencies
The "class reference" part has been omitted - passing a reference string
is much safer, in case callbacks are desired to be passed as a dependency
- or function references.

Also added symbol as a property to check for.

Thus, this method will only fail on "undefined" which in most cases would
cause issues anyways.